requirements.txt
tweepy==3.9.0
simplejson==3.17.2
python-dotenv==0.15.0